# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creditra API
  version: 0.1.0
  description: |
    Backend API for Creditra – on-chain credit scoring and risk evaluation.

    Browser clients are restricted by a CORS allowlist. Production deployments
    must set `CORS_ORIGINS` to a comma-separated list of exact origins; local
    development falls back to loopback origins when the variable is unset.

    ## Authentication
    Protected endpoints require a valid API key passed via the `X-API-Key` header.
    API keys are configured via the `API_KEYS` environment variable (comma-separated).

    ## Keeping the spec in sync
    - **Zod schemas** in `src/schemas/` are the runtime source of truth for request
      and response validation. This OpenAPI document should mirror those schemas.
    - Every new route **must** have a corresponding entry here before merging.
    - Run `npm run validate:spec` in CI to catch YAML/structural errors early.
    - Use path-level `operationId` values as the source of truth for client generation.
    - See `docs/json-schema-validation.md` for request/response validation policy.
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authenticated endpoints. Required for admin operations.
        Keys are configured via the `API_KEYS` environment variable.
  schemas:
...
    CreateCreditLineRequest:
      type: object
      required: [walletAddress, requestedLimit]
      properties:
        walletAddress:
          type: string
          minLength: 1
          maxLength: 256
          description: Stellar wallet address
          example: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
        requestedLimit:
          type: string
          pattern: "^\\d+(\\.\\d+)?$"
          description: Requested credit limit as a decimal string
          example: "1000.00"
      additionalProperties: false
...
    CreditLine:
      type: object
      properties:
        id:
          type: string
          description: Unique credit line identifier
          example: "cl_abc123"
        walletAddress:
          type: string
          description: Associated Stellar wallet address
          example: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
        creditLimit:
          type: string
          description: Approved credit limit
          example: "720.00000000"
        currentBalance:
          type: string
          description: Current outstanding balance
          example: "100.00000000"
        availableCredit:
          type: string
          description: Available credit (creditLimit - currentBalance)
          example: "620.00000000"
        interestRateBps:
          type: number
          description: Interest rate in basis points
          example: 640
        status:
          type: string
          enum: [active, suspended, closed]
          description: Current credit line status
        version:
          type: integer
          minimum: 1
          description: >
            Optimistic-locking version, incremented on each update. Pass the
            value you read back as `expectedVersion` on a subsequent update to
            guard against lost writes.
          example: 1
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 last update timestamp
    CreditLinesResponse:
      type: object
      required: [creditLines, pagination]
      properties:
        creditLines:
          type: array
          items:
            $ref: "#/components/schemas/CreditLine"
        pagination:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
          required: [total, offset, limit]
...
    RiskEvaluateRequest:
      type: object
      required: [walletAddress]
      properties:
        walletAddress:
          type: string
          description: Stellar wallet address to evaluate
          example: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
        forceRefresh:
          type: boolean
          description: Optional hint to bypass cache and recalculate risk
          default: false
      additionalProperties: false
    RiskEvaluateResponse:
      type: object
      required: [walletAddress, riskScore, creditLimit, interestRateBps, message]
      properties:
        walletAddress:
          type: string
          example: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
        riskScore:
          type: number
          description: Risk score (0-100)
          example: 72
        creditLimit:
          type: string
          description: Recommended credit limit
          example: "720.00000000"
        interestRateBps:
          type: number
          description: Recommended interest rate in basis points
          example: 640
        message:
          type: string
          example: "New risk evaluation completed"
    RiskEvaluateApiResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RiskEvaluateResponse"
        error:
          type: string
          nullable: true
    RiskEvaluation:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique evaluation identifier
        walletAddress:
          type: string
          description: Evaluated wallet address
        riskScore:
          type: number
          description: Risk score (0-100)
        creditLimit:
          type: string
          description: Recommended credit limit
        interestRateBps:
          type: number
          description: Recommended interest rate in basis points
        factors:
          type: object
          additionalProperties: true
          description: Risk evaluation factors and weights
        evaluatedAt:
          type: string
          format: date-time
          description: ISO 8601 evaluation timestamp
...
    ReconciliationTriggerResponse:
      type: object
      properties:
        data:
          type: object
          required: [jobId, message]
          properties:
            jobId:
              type: string
              description: Identifier for the scheduled reconciliation job
            message:
              type: string
              example: Reconciliation job scheduled
        error:
          type: string
          nullable: true
    ReconciliationStatusResponse:
      type: object
      properties:
        data:
          type: object
          required: [workerRunning, queueSize, failedJobs]
          properties:
            workerRunning:
              type: boolean
            queueSize:
              type: integer
              minimum: 0
            failedJobs:
              type: integer
              minimum: 0
        error:
          type: string
          nullable: true

    # ── Error Schemas ─────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [data, error]
      properties:
        data:
          nullable: true
          description: Always null on error
        error:
          type: string
        id:
          type: string
          description: Present on 404 responses that reference a specific resource
    ValidationErrorResponse:
      type: object
      required: [data, error, details]
      description: |
        Returned for Zod request-schema failures (body, query, or path params).
        Stable shape — never includes stack traces or internal exception text.
      properties:
        data:
          nullable: true
          example: null
        error:
          type: string
          enum: [Validation failed]
          example: Validation failed
        details:
          type: array
          items:
            type: object
            required: [field, message]
            additionalProperties: false
            properties:
              field:
                type: string
                description: Dotted path to the invalid field, or "(root)"
                example: walletAddress
              message:
                type: string
                example: walletAddress must be a valid Stellar address
      additionalProperties: false
    RateLimitErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Too many requests"
paths:
...
  /api/reconciliation/trigger:
    post:
      tags: [Reconciliation]
      operationId: triggerReconciliation
      summary: Schedule a reconciliation job
      description: Admin-gated endpoint that schedules a DB-vs-chain reconciliation job and returns immediately.
      responses:
        "202":
          description: Reconciliation job scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationTriggerResponse"
              example:
                data:
                  jobId: "reconciliation-1710000000000"
                  message: Reconciliation job scheduled
                error: null
        "401":
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to schedule reconciliation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/reconciliation/status:
    get:
      tags: [Reconciliation]
      operationId: getReconciliationStatus
      summary: Get reconciliation worker and queue status
      description: Admin-gated endpoint for dashboards and alerts to inspect worker and queue health.
      responses:
        "200":
          description: Reconciliation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationStatusResponse"
              example:
                data:
                  workerRunning: true
                  queueSize: 0
                  failedJobs: 0
                error: null
        "401":
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get reconciliation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Credit Lines Routes ─────────────────────────────────────────────────────
  /api/credit/lines:
    get:
      tags: [Credit]
      operationId: listCreditLines
      summary: List all credit lines
      security: []
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset (offset mode; ignored when `cursor` is present)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          description: Number of credit lines per page (validated by creditLinesQuerySchema)
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            When present (including empty string), switches to cursor pagination.
            Pass `nextCursor` from the previous page for subsequent pages.
      responses:
        "200":
          description: |
            Offset mode returns `{ data: { creditLines, pagination }, error: null }`.
            Cursor mode returns an unenveloped `{ creditLines, pagination }` body
            (see docs/cursor-pagination.md).
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed per window
              schema:
                type: integer
                example: 100
            X-RateLimit-Remaining:
              description: Requests remaining in current window
              schema:
                type: integer
                example: 95
            X-RateLimit-Reset:
              description: Unix timestamp when the rate limit window resets
              schema:
                type: integer
                example: 1700000060
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLinesResponse"
              example:
                data:
                  creditLines: []
                  pagination:
                    total: 0
                    offset: 0
                    limit: 100
                error: null
        "400":
          description: Invalid query parameter (Zod validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
                example: 47
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitErrorResponse"

    post:
      tags: [Credit]
      operationId: createCreditLine
      summary: Create a new credit line
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditLineRequest"
            example:
              walletAddress: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
              requestedLimit: "1000.00"
      responses:
        "201":
          description: Credit line created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLine"
        "400":
          description: Validation error or missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/risk/evaluate:
    post:
      tags: [Risk]
      operationId: evaluateRisk
      summary: Evaluate on-chain risk for a wallet
      description: |
        Performs a fresh risk evaluation for the given Stellar wallet address.
        Rate limited: default 10 requests per minute per IP address.
        Configure via RATE_LIMIT_MAX_EVALUATE and RATE_LIMIT_WINDOW_MS env vars.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskEvaluateRequest"
            example:
              walletAddress: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
              forceRefresh: false
      responses:
        "200":
          description: Risk evaluation result
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed per window
              schema:
                type: integer
                example: 10
            X-RateLimit-Remaining:
              description: Requests remaining in current window
              schema:
                type: integer
            X-RateLimit-Reset:
              description: Unix timestamp when the rate limit window resets
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvaluateApiResponse"
              example:
                data:
                  walletAddress: "GDRXE2BQUC3AZNPVFSCEZ76NJ3WWL25FYFK6RGZGIEKWE4SOUJ3LNLRK"
                  riskScore: 72
                  creditLimit: "720.00000000"
                  interestRateBps: 640
                  message: New risk evaluation completed
                error: null
        "400":
          description: Missing required field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Validation failed
        "413":
          description: Request body exceeds the 100 kb limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                data: null
                error: Request body too large. Maximum size is 100kb.
        "415":
          description: Content-Type is not application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                data: null
                error: Content-Type must be application/json
...
  /api/risk/evaluations/{id}:
    get:
      tags: [Risk]
      operationId: getRiskEvaluationById
      summary: Fetch a stored risk evaluation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Risk evaluation identifier
      responses:
        "200":
          description: Risk evaluation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvaluation"
        "404":
          description: Risk evaluation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Risk evaluation not found
                id: "abc123"
  /api/risk/wallet/{walletAddress}/latest:
    get:
      tags: [Risk]
      operationId: getLatestRiskEvaluation
      summary: Get the latest risk evaluation for a wallet
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
            pattern: '^G[A-Z2-7]{55}$'
          description: Stellar public key
      responses:
        "200":
          description: Risk evaluation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvaluateResponse"
        "400":
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No risk evaluation found for wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/risk/wallet/{walletAddress}/history:
    get:
      tags: [Risk]
      operationId: getRiskEvaluationHistory
      summary: Get risk evaluation history for a wallet
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
            pattern: '^G[A-Z2-7]{55}$'
          description: Stellar public key
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Pagination offset
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Pagination limit
      responses:
        "200":
          description: Array of risk evaluations
          content:
            application/json:
              schema:
                type: object
                properties:
                  evaluations:
                    type: array
                    items:
                      $ref: "#/components/schemas/RiskEvaluation"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"